+++ src/Adapter/Subject/ChoiceSubject.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./Subject.mjs").Subject} Subject */
+
+/**
+ * @typedef {{subjects: Subject[]}} ChoiceSubject
+ */